refactor(backend): move shared feature filters to common _backend module
Backend-hygiene fix (PR #110 follow-up review): SequenceFeature should not import
from NumericalFeature's dedicated backend package (_backend/num_feat/). Relocate
the shared model-free filter helpers into a common _backend/feature_filter.py that
both frontends import from.

- _backend/num_feat/filter_correlation.py -> _backend/feature_filter.py (git mv,
  history preserved); fold filter_variance_ into the same module.
- Remove _backend/num_feat/filter_variance.py (content moved).
- Update imports in _numerical_feature.py and _sequence_feature.py.
- No behavior change; prune + filter_correlation tests pass (38).
